Rework all operations to handle more scenarios
This expands the responsibilities of the language API to allow pulling
more (required) information about nodes. This is necessary in order to
handle 'wrapped' nodes like quoted lists in clojure: `'()`.

All operations are now more correct when dealing with a broader set of
operations and the implementation is also simpler to understand.