Add support for configurable cursor behaviour
Currently, when slurping or barfing, the cursor position always remains
in place. This can be annoying when during a barf the form is altered
such that the cursor now falls outside of its bounds. For example:

  `(aa |bb)` -> `(aa) |bb`

If barf is run again or the operation is dot-repeated then the barf
operation will no longer be applying to the same node.

To resolve this issue a customizable `cursor_behaviour` config option
has been added with three possible values:

+ `"remain"`
+ `"follow"`
+ `"auto"` [default]

When set to `auto`, if the cursor position falls out of bounds of the
original form, the cursor position will be reset to the moved bracket.

When set to `follow` the cursor position will always be moved to follow
the bracket.

When set to `remain` the cursor position will always remain in the same
position.