Fix motions which start from whitespace
When calling `get_node_at_cursor` from a whitespace char the parent or
outer node is returned. This is unexpected when performing a motion from
within a form.

This commit adds an additional check which returns the nearest adjacent
node within a form if a motion is triggered from a whitespace char.