Replace `operator` keymap field with smarter funcs
Currently the E/B motions are configured with `operator = true` which
additionally adds the `o` mode keymap with the fn wrapped in `normal! v`
to allow it to function in operator-pending mode.

This moves responsibility into the keymap setup phase instead of the
function handling the action, and in the current implementation won't
work in visual mode operator-pending states.

This commit shifts the responsibility to the motions API by allowing
them to detect the current mode and conditionally set `normal! v`.