Rework test command to be more flexible
This reworks the test command to work better when switching between
different nvim channels. Currently the .build/nvim dir needs to be
deleted manually when switching to a different channel but now different
channels are downloaded to separate directories.

Additionally this adds support for specifying an individual file to run
the tests for which improves the developer experience of working on
tests.