Revamp documentation
This current documentation was getting a bit out of date and out of
hand. Not everything was documented, some things were documented
incorrectly and everything was in a single file.

This updates the documentation to be up-to-date with the current state
of our API's and splits certain aspects into new files under `./docs`