Rewrite core on top of treesitter queries
The current core implementation of nvim-paredit relies on a language
extension api to ask questions about the AST.

This is not a very treesitter-like way of doing things and the extension
API was pretty gnarly.

This is a complete reimplementation which relies entirely on treesitter
queries with known capture groups. This drastically simplifies both the
implementation and the task of extending to other languages.

Now, instead of having to understand and write a language API, you can
just write a few treesitter queries to help nvim-paredit understand the
grammar.