Make document root assertions grammar agnostic
Currently the approach to checking if a node represents the document
root is to check if its `type` is `"source"`. While a lot of grammars do
this by convension it's not part of any spec and some grammars call it
something else - like `"program"`.

This replaces all these assertions with a comparison to see if the node
is :equal() to the tree :root() which is completely grammar agnostic.