Simplify utils for writing tests
The current utils for writing buffer operation tests are quite wordy and
not easy for a human to interpret at a glance.

In an effort to make it easier to write detailed, language specific
tests I have added a way to describe test prepare and check content as a
string.

This works by looking for `|` characters in test content and using this
to infer the cursor position.