Swap both elements of a pair together
Currently when performing a swap, each element in a pair is swapped with
it's corresponding paired element.

This works for simple cases but breaks when the number of lines are
different between the two pairs as the ranges the nodes represent have
now changed from previous edits.

This fixes the issue by finding the full range represented by both
elements in a pair and swapping this entire range in one go.